Исправляем сброс редактора изображения в дефолт. Рефакторинг модулей работы с модалками и редактором
